Choose protocol automatically
I have tried various error handling, this is probably the most effective in my opinion. looking at your previous code, the error callback is still not caught properly, for protocols that can't run by declaring at the beginning of the code, the nested function here aims to automatically select the protocol (if http can't be used, it automatically switches to https, if both errors, then it is returned ).

Thank you (^_^)